Further Cosmetic
Edit was improved binding the fields to the content of the actual
edditing word.
Remove word was improved, preventing the crash when no word was
selected.
New pages were added: WordQuiz page and in configuration: Edit Language
page, SortWordsPage.
The master page and pages changed color.

+++ lexiboom/lexiboom/View/ListofWords/ListofWords.xaml.cs
@@ -67,12 +67,19 @@ namespace lexiboom.View.ListofWords
 
         async void OnRemoveClicked(object sender, EventArgs e)
         {
-            bool answer = await DisplayAlert("Remove word?", "Are you sure you want to get rid of this word?", "Yes", "No");
-            if(answer)
+            if (listView.SelectedItem == null)
             {
-                await App.Storage.connection.DeleteAsync(listView.SelectedItem);
-                App.Storage.listofWordsList.Remove((MotherTongeWords)listView.SelectedItem);
-            }            
+                await DisplayAlert("Error", "No row was selected to be removed, did you miss something?", "Ok");
+            }
+            else
+            {
+                bool answer = await DisplayAlert("Remove word?", "Are you sure you want to get rid of this word?", "Yes", "No");
+                if (answer)
+                {
+                    await App.Storage.connection.DeleteAsync(listView.SelectedItem);
+                    App.Storage.listofWordsList.Remove((MotherTongeWords)listView.SelectedItem);
+                }
+            }                        
         }
 
         public async void TextCell_Tapped(object sender, EventArgs e)
